requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
lxml==4.5.2
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
vk-api==11.8.0
Pillow==7.2.0
peewee==3.13.3
peewee-migrations==0.3.19
